removed the interpolation of network outputs and reduced 1 step. Added interpolation to the alb url
